chore: refactor col ratio functions
